Creating JDBC sink to persist message payloads to a database table
Transforms JSON payload to a map with specified column values to provide
data values for insert statment, defaults to single column named payload
for entire payload content as is.

Defaults to reading database properties from config/jdbc.properties

Unless these properties are modified the database used is an in-memory
embedded HSQLDB instance.

This sink can be configured with the following options:

- configProperties (base name of properties file, default is jdbc)

- initializeDatabase (default is false)
  (the default jdbc.properties has this set to true)

- initializerScript (default is init_db.sql)
  (will search config directory for this file)

- columns (comma separated list of column names to include)

- the usual JDBC properties - driverClass, url, username, password