XD-792 Use SI twitter search support.
TwitterSearchMessage source was a temporary fix to provide
authenticated twitter search until it was supported by Spring
Integration.

However, we still need a customised versions of TwitterTemplate
which can be configured with the consumerKey and consumerSecret as
Spring Social doesn't provide a matching constructor which will obtain
an OAuth2 bearer token internally.