XD-988: Batch export from HDFS to JDBC.
Note the issue with Spring HADOOP (SHDP-201) which prevents the easy use
of both HDFS and file resources in the same module, thus preventing the
use of the same DB initialization code from the file->JBDC module. This
is commented out for now and relies on manually pre-creating the
database table.