XD-1186: Support for overriding application.yml
Use Spring Boot's spring.config.location option to point to an additional
configuration file which can be used to override the defaults.

Introduces XD_CONFIG environment variable which defaults to
config/xd-config.yml in the startup scripts but which a user can set to
point to a filesystem location outside XD if they wish.

The file config/xd-config.yml is commented out by default so has no effect
unless edited.