XD-1194 Handle boolean cmdLineOption default
Fixed 'jmxEnabled' CommonOption to use Boolean type instead of boolean
so that getter method could return null when the flag is not set and,
the setter would always set the appropriate precedence value when
resolving PropertySources.
Also note we have the application.yml setting XD_JMX_ENABLED to false hence
setter is always set with non-null value.

XD-1194 Remove null check on Boolean type

The getter method on XD_JMX_ENABLED in CommonOptions can never return false
as it can be either true or null.
Hence, made it to return value as is.