XD-1246 Run tests with existing admin/launcher
- Added RandomConfigurationSupport class that sets random config
values for:
1) Admin server port
2) HSQLDB host, port, database, dbname and data directory
3) XD deployer/undeployer name
These config values are set as system properties so that they get the
highest precedence next to commandLine properties.

- Added proeprty-placeholder configurer for "xd.deployer" and "xd.undeployer"
that defaults to existing names "xd.deployer" and "xd.undeployer". This will
enable us to override the names during tests so that the deployment requests
will be sent to the configured queues/topics/exchanges (depending on the bus)

- Fixed the tests that starts adminServer/container by adding RandomConfigurationSupport
- Fixed the xd.deployer name in ModuleDeploymentTests
- Fixed eclipse metadata file for license file to read year 2014

XD-1246 Fix review comments

- Renamed the queue/topic placeholder names
- Updated commons-io to use version variable which sets to '2.4'