Fix NPE on SingleNodeStreamDeployment test
- Changed the AbstractSingleNodeStreamDeploymentIntegrationTests' shutdown
method to be called after the cleanup method in MessageBus specific integration
tests. Since the cleanup methods on those classes use the context, this will
always make sure the context is available.
  - Created junit ExternalResource @ClassRule and implemented after() method
to call the shutdown so that it will be invoked after the @AfterClass based on
junit lifecycle methods.