Fix broken twitter modules.
Stream and search were both broken due to

- Changes in version of httpclient expected by Spring
- Inconsistent jackson versions between dirt and module

This modifies the reflection hack to account for both types of client
connection factory which Spring can create. Currently adding httpclient
4.3.3 causes a separate problem where the build hangs when running with
rabbit, so the original plan of ensuring httpclient was on the classpath
and therefore an HttpComponentsClientHttpRequestFactory was created is
not acceptable.