XD-1357
If ZooKeeper is available based on the client connect string, the Container's host and PID are written to an ephemeral node under /xd/containers.

That node's name will match the Container ID.

The ID is generated within ContainerMetadata and set on the ApplicationContext via an initializer.

Removed the ContainerStartedEvent publishing from ContainerServerApplication.