XD-1372 Aligns Spring XD Reactor to Reactor ver1.1
Additional changes include:

- Incorporated an InboundChannelAdapter based on the NetServerFactoryBean which allows for a simplifed configuration of a TCP or UDP server using the Reactor NetServer abstra
- Updated the XML namespace to include the NetServerInboundChannelAdapter configuration and created a parser for the "netserver-inbound-channel-adapter" element.
- Added a test for the new NetServerInboundChannelAdapter
- Added a batching MessageHandler that uses the RingBuffer directly without built-in, other-thread event handling (all events occur in the calling thread). This implementatio
- Since a MessageChannel implementation is now included in the reactor-spring-messaging artifact, the one from this location was deleted.

XD-1372 Update to reflect changes suggested on the PR #625

XD-1372 Update to reflect changes suggested on the PR #625

Completely removed the netserver namespace as the configuration was uncessarily complicated. Now relying on @Configuration JavaConfig to configure the necessary components.

To that end, created a NetServerSpecFactoryBean and changed the NetServerInboundChannelAdapter to accept the spec, which is the Reactor component holding all the pre-start configuration. The FactoryBean now has the logic of parsing a URI and figuring out what components to configure where. If creating a NetServerInboundChannelAdapter in JavaConfig and needing more flexibility than the string-based property-oriented configuration method allows, simply pass in a configured NetServerSpec with whatever options you wish (including ServerSocketOptions like tcpnodelay() reuseAddress() and so on)

Also hacked the tests to simulate an XD module environment by providing a shim between the ModuleOptionsMetadata class and the StringValueResolver needed to properly bind external values into the base @Configuration class.

XD-1372 Tweak to options metadata description PR #625

Polishing

* Reformat to XD rules
* Move `@EnableIntegration` to the base configuration class
* Remove unused env field from factory bean