XD-1372 Update to reflect changes suggested on the PR #625
Completely removed the netserver namespace as the configuration was uncessarily complicated. Now relying on @Configuration JavaConfig to configure the necessary components.

To that end, created a NetServerSpecFactoryBean and changed the NetServerInboundChannelAdapter to accept the spec, which is the Reactor component holding all the pre-start configuration. The FactoryBean now has the logic of parsing a URI and figuring out what components to configure where. If creating a NetServerInboundChannelAdapter in JavaConfig and needing more flexibility than the string-based property-oriented configuration method allows, simply pass in a configured NetServerSpec with whatever options you wish (including ServerSocketOptions like tcpnodelay() reuseAddress() and so on)

Also hacked the tests to simulate an XD module environment by providing a shim between the ModuleOptionsMetadata class and the StringValueResolver needed to properly bind external values into the base @Configuration class.