Refactored code to support Http as a source.
Had to move HttpSource back to XD-Shell project.  There are tests in the XD-Shell that explicitly test the http post command and require that the HTTPSource be present.  I created SimpleHttpSource for acceptance tests that sends http posts without using any of the XD Shell utils.

I went ahead and updated tests that were using http as source (except http test) to use trigger.  This helped improve testing speed, because we don't have to use waitForXD's.