XD-1456 Using Spring Boot as base for Acc tests.
The previous version of Acceptance tests was driven by environment variables to configure
the tests and this became unwieldy.   This PR setup acceptance tests to use @SpringApplicationConfiguration and XDEnvironment to configure the tests. The other benefit is that through the use of profiles a user can setup test scenarios as different profiles.

Other things done:
* Added MQTT Test
* Foundation for JMS test (JMS Source ) has been added, but left out the test because it already has its own Jira.
* Created singlenode and ec2 properties for users.
* Removed all test configuration properties from build.gradle
* Instructions now reflect the use of SpringApplicationConfiguration