Provided names for PathChildrenCache threads
When reading a thread dump or stack trace generated
by a PathChildrenCache thread it is helpful to
know which cache that thread is serving. This
change provides a meaningful name to the thread
Curator creates for these caches.

Also introduced an AtomicReference for the containers
cache. The idea is that the reference isn't set
until the cache is started and the reference is
removed before the cache is closed.