Fixed issue with stream/job undeployment
When undeploying a stream/job the stream/job definition is loaded
in order to perform a graceful undeployment of the underlying
modules. However if the stream/job definition has already
been removed from /xd/streams|/xd/jobs then the definition
cannot be loaded. This fix will prevent an attempt to read
a definition that has been removed.