XD-1516: Deploy "orphaned" jobs to new containers
When a new container starts, determine if there are any
deployed jobs that currently are not running on a container;
if so then deploy them to the new container.

There is some copy/paste code here (see comment in
ContainerListener line 179, method onChildAdded)
that needs to be refactored later on.