XD-1540: Stacktrace on xd-admin restart
Fixed race condition between StreamListener and ContainerListener.

Upon buildup of the stream deployment cache, the listener fires
which causes the streams to be deployed. At around the same time,
the container shows up which causes its listener to try to deploy
a stream to it. The stream listener was the first to write
the deployment path thus causing the subsequent NodeExists exception
in ContainerListener.