Renamed classes
JobListener and StreamListener were renamed to
JobDeploymentListener and StreamDeploymentListener to
more accurately reflect their roles. AdminServer was
renamed to DeploymentSupervisor since this class
is not actually a server in a traditional sense
(i.e. it does not directly respond to requests) but
instead is the single process in the system that
is responsible for managing containers and deployments.