XD-1389 Partitioned Batch Improvements
JIRA: https://jira.spring.io/browse/XD-1389

Occasional NPE in ftphdfs job.

There was a fixed timeout of 60 seconds for the partition results
to be returned to master.

The spring-batch `MessageChannelPartitionHandler` croaks with
an NPE on timeout - this has now been corrected on master (batch).

In the config import file, Increase the default timeout to 1 hour
and make it configurable using the `partitionResultsTimeout` property.

In the ftphdfs job properties, make the default 5 minutes.

Also, remove `stepConcurrency`. Step concurrency will be handled
by the upcoming partitioning changes (in the message bus) XD-1338.