XD-1670: NPE when a container departs
When a container departs the cluster an attempt is made to
redeploy any modules it was running. However the stream
may have been destroyed or undeployed by the time the container's
ephemeral node goes away.

This fix will ignore modules for streams that don't exist
or aren't deployed.

Jobs aren't handled in this fix; they will be addressed
as part of https://jira.spring.io/browse/XD-1548.