XD-1575
Fixes XD-1575 by adding support for UDP to the SyslogInboundChannelAdapter. Internally it forks on the Tcp/Datagram configuration depending on the value of the "transport" configuration option. The test was updated to run the TCP tests by default, since they are the only ones that can be run reliably on CI servers. The same tests can be run in UDP mode by specifying -Dtransport=udp on the command line when running the tests but they must be run manually because the tests assume that all messages reach the destination which is not a guarantee with UDP so there's no way to reliably test whether messages were properly sent or not.

XD-1575 updates Reactor to version 1.1.0.RELEASE

XD-1575 minor tweaks