XD-1739: Container reconnection to ZK fails intermittently
If a container disconnects and reconnects to ZooKeeper, sometimes
the ephemeral node for that container isn't cleaned up before
the container attempts to create a new ephemeral node. This fix
will delete the node if it exists prior to creating a new one.

In theory this should never happen, but in cases where the ZK
server and the container are both suspended (i.e. a laptop goes
to sleep) the container could reconnect before ZK cleans up
the old ephemeral node.

Also added the @SuppressWarnings("serial") annotation to
ZooKeeperAccessException.