XD-1338 Hints for partitioning a stream (cont)
This is a continuation of XD-1338. It includes the following:

When writing a module deployment request, the module deployment
properties are written as the node data. These properties
are read by the container executing the deployment.

When a container leaves the cluster, those properties are reused
by the container that the module is redeployed to. This allows
the module deployment to carry forward partition information
(especially partition index) that is unique to that deployment.

Introduced ContainerAwareModuleDeploymentPropertiesProvider to provide
deployment properties for modules targeted to a specific container.
It is an extension of ModuleDeploymentPropertiesProvider that
can be used in areas where deployment properties for a specific
container are required.

Also renamed ExcludingContainerMatcher to RedeploymentContainerMatcher
and modified the implementation to only ever return one Container.

A typo in MessageBusSupport was corrected.

Addressed comments for PR 902