XD-1783: Exceptions thrown on module redeployment
As part of the ZK refactoring, the logging we used to do
when there were no containers for deployment was lost.
The exception NoContainerException was introduced
to indicate that a deployment failed because there were
no containers that met the criteria for deployment.

This is a checked exception since this condition needs
to be handled whenever a module deployment is attempted.